relocate service and interface to appropriate folders. update corresponding namespaces.
