+++ src/main/Plugins/ValidationPlugins/Dns/Script/Script.cs
@@ -2,6 +2,7 @@
 using PKISharp.WACS.Clients;
 using PKISharp.WACS.DomainObjects;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Interfaces;
 
 namespace PKISharp.WACS.Plugins.ValidationPlugins.Dns
 {
