+++ src/plugin.validation.dns.dreamhost/DreamhostDnsValidation.cs
@@ -1,7 +1,7 @@
 ﻿using Nager.PublicSuffix;
-using PKISharp.WACS.Plugins.ValidationPlugins.Dns;
 using PKISharp.WACS.Plugins.ValidationPlugins.Dreamhost;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Interfaces;
 
 namespace PKISharp.WACS.Plugins.ValidationPlugins
 {
